# Store tables as PARQUET files