… audformat/core/utils.py
… pyproject.toml
… tests/test_database.py
… tests/test_misc_table.py
… tests/test_table.py