@hagenw 2024-06-19
> The hash key's value seems kind of generic but I wonder whether it only had been introduced recently, or whether it is changed sometimes?

This is indeed a good point. I created https://github.com/audeering/audformat/pull/436 to propose a better hash solution.