@hagenw 2024-06-19
Regarding the complexity of `audformat.Table._pyarrow_convert_dtypes()`. If you have correctly spotted, the conversion can only be done if the table is assigned to a database, and hence has information on the used schemes and columns.
Originally, I thought I might be able to simplify it by adding `index_names`, `index_dtypes`, `column_names`, `column_dtypes` as arguments, but the problem is that it also needs to inspect `self.db.schemes`.

So far, I have not found a better way to restructure the code, but I agree that we should make `audformat.Table._pyarrow_convert_dtypes()` a class method and add unit tests for it. I created https://github.com/audeering/audformat/issues/437 to track this.